// adapters/implementations/codex-app-server/src/protocol/generated/v2/ThreadResumeResponse.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ReasoningEffort } from '../ReasoningEffort';
import type { AskForApproval } from './AskForApproval';
import type { SandboxPolicy } from './SandboxPolicy';
import type { Thread } from './Thread';

export type ThreadResumeResponse = {
  thread: Thread;
  model: string;
  modelProvider: string;
  cwd: string;
  approvalPolicy: AskForApproval;
  sandbox: SandboxPolicy;
  reasoningEffort: ReasoningEffort | null;
};